separate beam layout from beam element
