+++ src/write/layout/calcAverage.js
@@ -1,0 +1,8 @@
+function calcAverage(total, numElements) {
+	if (!numElements)
+		return 0;
+	return total / numElements;
+}
+
+module.exports = calcAverage;
+
