+++ src/write/layout/getBarYAt.js
@@ -1,0 +1,6 @@
+function getBarYAt(startx, starty, endx, endy, x) {
+	return starty + (endy - starty) / (endx - startx) * (x - startx);
+}
+
+module.exports = getBarYAt;
+
